refactor(omegaform): create localized validation hooks + extract per-field generator
